Redo previous fixes (some buggy) properly JIRA: CXX-5164
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@64137 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/test/test_ncbi_pipe.cpp
@@ -208,11 +208,10 @@ static void s_WriteLine(FILE* fs, const string& str)
 static void s_ReadStream(istream& ios)
 {
     size_t total = 0;
-
     for (;;) {
         char   buf[kBufferSize];
         ios.read(buf, sizeof(buf));
-        size_t cnt = ios.gcount();
+        size_t cnt = (size_t) ios.gcount();
         ERR_POST(Info << cnt << " byte(s) read from stream"+string(":"+!cnt));
         if (cnt) {
             NcbiCerr.write(buf, cnt);
